break up types and styles into new files
